Added background script to resolve the CORS issue on chrome. Issue - #9
