+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "Comics / Manga Downloader",
-  "version": "1.2.0",
+  "version": "1.2.1",
   "description": "A cross browser extension to download comics / manga as pdf file from any site.",
   "author": "Saptarshi Mondal <mondalsaptarshi7@gmail.com>",
   "license": "MIT",
