+++ src/popup/index.html
@@ -330,7 +330,7 @@
               <button id="clearQuery" class="" title="Toggle advanced filters">
                 <img class="toggle" src="./images/times.svg">
               </button>
-              <button id="open_options_button" title="Options">
+              <button id="open_options_button" title="Options" style="visibility: hidden;">
                 <img src="./images/cog.svg"></button>
               </div>
             <div>
